Refactor commands ingest module by splitting functionalities into separate files
- Introduced `commit_artefacts.rs` to handle commit artefact appending logic.
- Introduced `current_mirror.rs` to manage mirroring of current artefacts.
- Updated `orchestrator.rs` to utilize the new modules and removed redundant code.
