+++ bitloops/src/daemon/process/spawn.rs
@@ -1,0 +1,24 @@
+use super::*;
+
+pub(in crate::daemon) fn current_binary_fingerprint() -> Result<String> {
+    let current_exe = env::current_exe().context("resolving Bitloops executable path")?;
+    let bytes = fs::read(&current_exe)
+        .with_context(|| format!("reading Bitloops executable {}", current_exe.display()))?;
+    Ok(hex::encode(Sha256::digest(bytes)))
+}
+
+pub(in crate::daemon) fn build_daemon_spawn_command(
+    args: &InternalDaemonProcessArgs,
+) -> Result<Command> {
+    #[cfg(unix)]
+    use std::os::unix::process::CommandExt;
+
+    let executable = env::current_exe().context("resolving Bitloops executable for daemon")?;
+    let mut command = Command::new(executable);
+    command.args(args.argv());
+    #[cfg(unix)]
+    {
+        command.process_group(0);
+    }
+    Ok(command)
+}
